refactor: FormCheckActivity 파일 이동
